refactored out code/assets
