+++ assets/loc_id_map.json
