+++ assets/models.js
@@ -1,0 +1,13 @@
+/**
+ * @file Colors and names of models used in viz
+ */
+
+const models = [
+  { name: 'Delphi', color: '#e84118', active: true },
+  { name: 'IHME_MS_SEIR', color: '#44bd32', active: true },
+  { name: 'Imperial', color: '#8c7ae6', active: true },
+  { name: 'LANL', color: '#0097e6', active: true },
+  { name: 'SIKJalpha', color: '#e1b12c', active: true },
+];
+
+export default models;
